Fix graceful shutdown to be more selective about when it sleeps
The server sleeps as part of graceful shutdown, which is to allow time
for load balancers to stop sending requests to the Iglu Server.

The server might need to shut down for several reasons, including
receiving a SIGTERM or because of an internal error. After this PR, the
server does the sleep only if shutting down for SIGTERM; it does not
sleep if shutting down for any other reason.